+++ app/src/main/java/com/codesoom/assignment/controllers/product/ProductController.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.controllers.product;
 
 import com.codesoom.assignment.application.ProductService;
 import com.codesoom.assignment.domain.product.Product;
